Move TopNSequence to druid-core.
